requirements.txt
# Face Recognition Attendance System - Core Dependencies
# Python 3.12+ compatible (Pins removed for build success)

# Computer Vision & Image Processing
opencv-contrib-python
Pillow

# Face Detection
mtcnn

# Face Recognition & Deep Learning
deepface
tensorflow
keras

# Data Processing
numpy
pandas

# Database
sqlalchemy
alembic

# API Framework
fastapi
uvicorn
python-multipart
pydantic
pydantic-settings

# Admin Dashboard
streamlit
plotly

# Security & Authentication
python-jose[cryptography]
passlib[bcrypt]
python-dotenv

# Utilities
scipy
python-dateutil
pytz
tqdm
requests

# Testing
pytest
pytest-cov
httpx

# GPU Acceleration & Advanced Models
# PyTorch with CUDA support (install separately via pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118)
# torch - Install via: pip install torch torchvision torchaudio --index-url https://download.pytorch.org/whl/cu118
# torchvision
# torchaudio

# ONNX Runtime with GPU
onnxruntime-gpu

# Super-Resolution & Enhancement
gfpgan
realesrgan
basicsr

# Additional GPU utilities
nvidia-ml-py3
